LPS-66054 Test CKEditorConfigContributor logic too
