+++ source-formatter.properties
@@ -125,7 +125,7 @@
     #
     line.length.excludes=\
         modules/apps/collaboration/item-selector/item-selector-editor-configuration/src/main/java/com/liferay/item/selector/editor/configuration/BaseEditorConfigContributor.java@34,\
-        modules/apps/collaboration/wiki/wiki-editor-configuration/src/test/java/com/liferay/wiki/editor/configuration/WikiAttachmentAlloyEditorEditorConfigContributorTest.java,\
+        modules/apps/collaboration/wiki/wiki-editor-configuration/src/test/java/com/liferay/wiki/editor/configuration/WikiAttachmentCKEditorEditorConfigContributorTest.java,\
         modules/apps/collaboration/wiki/wiki-test/src/testIntegration/java/com/liferay/wiki/subscription/test/WikiSubscriptionContainerModelTest.java,\
         modules/apps/collaboration/wiki/wiki-test/src/testIntegration/java/com/liferay/wiki/trash/test/WikiPageDependentsTrashHandlerTest.java,\
         modules/apps/collaboration/wiki/wiki-test/src/testIntegration/java/com/liferay/wiki/trash/test/WikiPageTrashHandlerTest.java@185,\
